requirements.txt
feedparser==6.0.12
beautifulsoup4==4.14.3
requests==2.32.5
PyYAML==6.0.1
